reorganized domain model classes in a server/model package and added a server/controllers package
