+++ app/src/main/java/cs_3560_project/server/model/Author.java
@@ -1,4 +1,4 @@
-package cs_3560_project.server;
+package cs_3560_project.server.model;
 
 import java.util.List;
 import jakarta.persistence.CascadeType;
