+++ app/src/main/java/cs_3560_project/server/model/Book.java
@@ -1,4 +1,4 @@
-package cs_3560_project.server;
+package cs_3560_project.server.model;
 
 import java.time.LocalDate;
 import java.util.List;
