+++ app/src/main/java/cs_3560_project/server/model/Documentary.java
@@ -1,4 +1,4 @@
-package cs_3560_project.server;
+package cs_3560_project.server.model;
 
 import java.time.LocalDate;
 import jakarta.persistence.CascadeType;
