+++ app/src/main/java/cs_3560_project/server/model/Person.java
@@ -1,0 +1,25 @@
+package cs_3560_project.server.model;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.MappedSuperclass;
+
+@MappedSuperclass
+public class Person {
+
+  @Column(name = "name")
+  private String name;
+
+  public Person() {};
+
+  public Person(String name) {
+    this.name = name;
+  }
+
+  public String getName() {
+    return name;
+  }
+
+  public void setName(String name) {
+    this.name = name;
+  }
+}
